fix(statistics fetch): fix it with the good structure
